+++ lib/pluck_in_batches/iterator.rb
@@ -2,12 +2,15 @@
 
 module PluckInBatches
   class Iterator # :nodoc:
+    VALID_ORDERS = [:asc, :desc].freeze
+    DEFAULT_ORDER = :asc
+
     def initialize(relation)
       @relation = relation
       @klass = relation.klass
     end
 
-    def each(*columns, start: nil, finish: nil, batch_size: 1000, error_on_ignore: nil, cursor_column: @relation.primary_key, order: :asc, &block)
+    def each(*columns, start: nil, finish: nil, batch_size: 1000, error_on_ignore: nil, cursor_column: @relation.primary_key, order: DEFAULT_ORDER, &block)
       if columns.empty?
         raise ArgumentError, "Call `pluck_each' with at least one column."
       end
@@ -18,18 +21,18 @@ module PluckInBatches
         end
       else
         enum_for(__callee__, *columns, start: start, finish: finish, batch_size: batch_size, error_on_ignore: error_on_ignore, cursor_column: cursor_column, order: order) do
-          apply_limits(@relation, start, finish, order).size
+          apply_limits(@relation, start, finish, build_batch_orders(order)).size
         end
       end
     end
 
-    def each_batch(*columns, start: nil, finish: nil, batch_size: 1000, error_on_ignore: nil, cursor_column: @relation.primary_key, order: :asc)
+    def each_batch(*columns, start: nil, finish: nil, batch_size: 1000, error_on_ignore: nil, cursor_column: @relation.primary_key, order: DEFAULT_ORDER)
       if columns.empty?
         raise ArgumentError, "Call `pluck_in_batches' with at least one column."
       end
 
-      unless order == :asc || order == :desc
-        raise ArgumentError, ":order must be :asc or :desc, got #{order.inspect}"
+      unless Array(order).all? { |ord| VALID_ORDERS.include?(ord) }
+        raise ArgumentError, ":order must be :asc or :desc or an array consisting of :asc or :desc, got #{order.inspect}"
       end
 
       pluck_columns = columns.map(&:to_s)
@@ -44,10 +47,11 @@ module PluckInBatches
       end
 
       relation = @relation
+      batch_orders = build_batch_orders(cursor_columns, order)
 
       unless block_given?
         return to_enum(__callee__, *columns, start: start, finish: finish, batch_size: batch_size, error_on_ignore: error_on_ignore, cursor_column: cursor_column, order: order) do
-          total = apply_limits(relation, cursor_columns, start, finish, order).size
+          total = apply_limits(relation, cursor_columns, start, finish, batch_orders).size
           (total - 1).div(batch_size) + 1
         end
       end
@@ -62,8 +66,8 @@ module PluckInBatches
         batch_limit = remaining if remaining < batch_limit
       end
 
-      relation = relation.reorder(*batch_order(cursor_columns, order)).limit(batch_limit)
-      relation = apply_limits(relation, cursor_columns, start, finish, order)
+      relation = relation.reorder(batch_orders.to_h).limit(batch_limit)
+      relation = apply_limits(relation, cursor_columns, start, finish, batch_orders)
       relation.skip_query_cache! # Retaining the results in the query cache would undermine the point of batching
       batch_relation = relation
 
@@ -99,9 +103,13 @@ module PluckInBatches
           end
         end
 
-        batch_relation = batch_condition(
-          relation, cursor_columns, cursor_column_offsets, order == :desc ? :lt : :gt
-        )
+        _last_column, last_order = batch_orders.last
+        operators = batch_orders.map do |_column, order| # rubocop:disable Lint/ShadowingOuterLocalVariable
+          order == :desc ? :lteq : :gteq
+        end
+        operators[-1] = (last_order == :desc ? :lt : :gt)
+
+        batch_relation = batch_condition(relation, cursor_columns, cursor_column_offsets, operators)
       end
     end
 
@@ -135,29 +143,33 @@ module PluckInBatches
         end
       end
 
-      def apply_limits(relation, columns, start, finish, order)
-        relation = apply_start_limit(relation, columns, start, order) if start
-        relation = apply_finish_limit(relation, columns, finish, order) if finish
+      def apply_limits(relation, columns, start, finish, batch_orders)
+        relation = apply_start_limit(relation, columns, start, batch_orders) if start
+        relation = apply_finish_limit(relation, columns, finish, batch_orders) if finish
         relation
       end
 
-      def apply_start_limit(relation, columns, start, order)
-        batch_condition(relation, columns, start, order == :desc ? :lteq : :gteq)
+      def apply_start_limit(relation, columns, start, batch_orders)
+        operators = batch_orders.map do |_column, order|
+          order == :desc ? :lteq : :gteq
+        end
+        batch_condition(relation, columns, start, operators)
       end
 
-      def apply_finish_limit(relation, columns, finish, order)
-        batch_condition(relation, columns, finish, order == :desc ? :gteq : :lteq)
+      def apply_finish_limit(relation, columns, finish, batch_orders)
+        operators = batch_orders.map do |_column, order|
+          order == :desc ? :gteq : :lteq
+        end
+        batch_condition(relation, columns, finish, operators)
       end
 
-      def batch_condition(relation, columns, values, operator)
-        columns = Array(columns)
-        values = Array(values)
-        cursor_positions = columns.zip(values)
+      def batch_condition(relation, columns, values, operators)
+        cursor_positions = Array(columns).zip(Array(values), operators)
 
-        first_clause_column, first_clause_value = cursor_positions.pop
+        first_clause_column, first_clause_value, operator = cursor_positions.pop
         where_clause = build_attribute_predicate(first_clause_column, first_clause_value, operator)
 
-        cursor_positions.reverse_each do |column_name, value|
+        cursor_positions.reverse_each do |column_name, value, operator| # rubocop:disable Lint/ShadowingOuterLocalVariable
           where_clause = build_attribute_predicate(column_name, value, operator == :lteq ? :lt : :gt).or(
             build_attribute_predicate(column_name, value, :eq).and(where_clause)
           )
@@ -170,9 +182,9 @@ module PluckInBatches
         @relation.bind_attribute(column, value) { |attr, bind| attr.public_send(operator, bind) }
       end
 
-      def batch_order(cursor_columns, order)
-        cursor_columns.map do |column|
-          @relation.arel_table[column].public_send(order)
+      def build_batch_orders(cursor_columns, order)
+        cursor_columns.zip(Array(order)).map do |column, ord|
+          [column, ord || DEFAULT_ORDER]
         end
       end